refactor(General): :recycle: Refactor mock data
