+++ cypress/e2e/skupper-console/sites/site.cy.ts
@@ -1,19 +1,26 @@
 import { getTestsIds } from '../../../../src/config/testIds';
+import sitesData from '../../../../mocks/data/SITES.json';
+import processesData from '../../../../mocks/data/PROCESSES.json';
+
+const site = sitesData.results[0];
+const sitePath = `#/sites/${encodeURIComponent(site.name)}@${encodeURIComponent(site.identity)}`;
+const process = processesData.results[0];
+const processPath = `#/processes/${process.name}@${process.identity}`;
 
 context('Site Detail', () => {
-  const testId = `${getTestsIds.siteView('site-1')}`;
+  const testId = `${getTestsIds.siteView(site.identity)}`;
 
   beforeEach(() => {
-    cy.visit('#/sites/site%201@site-1?type=Details');
+    cy.visit(`${sitePath}?type=Details`);
   });
 
   it('should navigate to the Site section', () => {
-    cy.get(`[data-testid="${testId}"]`).contains('site 1').should('have.length', 1);
-    cy.get(`[data-testid="${testId}"]`).contains('config-grpc-site-1-test').should('have.length', 1);
+    cy.get(`[data-testid="${testId}"]`).contains(site.name).should('have.length', 1);
+    cy.get(`[data-testid="${testId}"]`).contains(site.nameSpace).should('have.length', 1);
   });
 
   it('should navigate to the correct site page when clicking on a link', () => {
-    const expectedUrl = 'http://localhost:3000/#/sites/site%201@site-1?type=Overview';
+    const expectedUrl = `http://localhost:3000/${sitePath}?type=Overview`;
 
     const link = cy.get(`[data-testid="${testId}"]`).contains('Overview');
     link.click();
@@ -22,17 +29,17 @@ context('Site Detail', () => {
 });
 
 context('Site Processes', () => {
-  const testId = `${getTestsIds.siteView('site-1')}`;
+  const testId = `${getTestsIds.siteView(site.identity)}`;
 
   beforeEach(() => {
-    cy.visit('#/sites/site%201@site-1?type=Processes');
+    cy.visit(`${sitePath}?type=Processes`);
   });
 
   it('should navigate to the correct process page when clicking on a link', () => {
     const expectedUrl = 'http://localhost:3000/#/processes/process%20payment%201@process-payment-1?type=Overview';
 
-    const link = cy.get(`[data-testid="${testId}"]`).contains('process payment 1');
-    link.should('have.attr', 'href', `#/processes/process payment 1@process-payment-1`);
+    const link = cy.get(`[data-testid="${testId}"]`).contains(process.name);
+    link.should('have.attr', 'href', `${processPath}`);
     link.click();
     cy.url().should('eq', expectedUrl);
   });
