+++ cypress/e2e/skupper-console/sites/sites.cy.ts
@@ -1,3 +1,5 @@
+import sitesData from '../../../../mocks/data/SITES.json';
+
 context('Sites', () => {
   beforeEach(() => {
     cy.visit('#/sites');
@@ -5,10 +7,10 @@ context('Sites', () => {
 
   it('should navigate to the Sites section', () => {
     cy.get('[data-testid=sk-sites-view]').contains('Sites').should('have.length', 1);
-    cy.get('[data-testid=sk-sites-view]').contains('site 1').should('have.length', 1);
-    cy.get('[data-testid=sk-sites-view]').contains('site 2').should('have.length', 1);
-    cy.get('[data-testid=sk-sites-view]').contains('site 3').should('have.length', 1);
-    cy.get('[data-testid=sk-sites-view]').contains('site 4').should('have.length', 1);
-    cy.get('[data-testid=sk-sites-view]').contains('site 5').should('have.length', 1);
+    cy.get('[data-testid=sk-sites-view]').contains(sitesData.results[0].name).should('have.length', 1);
+    cy.get('[data-testid=sk-sites-view]').contains(sitesData.results[1].name).should('have.length', 1);
+    cy.get('[data-testid=sk-sites-view]').contains(sitesData.results[2].name).should('have.length', 1);
+    cy.get('[data-testid=sk-sites-view]').contains(sitesData.results[3].name).should('have.length', 1);
+    cy.get('[data-testid=sk-sites-view]').contains(sitesData.results[4].name).should('have.length', 1);
   });
 });
