+++ cypress/tsconfig.json
@@ -4,7 +4,9 @@
     "target": "ES2022",
     "lib": ["dom", "es2022"],
     "types": ["cypress"],
-    "baseUrl": "../src"
+    "baseUrl": "../src",
+    "resolveJsonModule": true,
+    "moduleResolution": "node"
   },
   "include": ["**/*.ts"]
 }
