+++ mocks/data/PROCESS_PAIRS.json
@@ -2,7 +2,7 @@
   "results": [
     {
       "identity": "process-catalog-to-process-cart-1",
-      "startTime": 1674202485242353,
+      "startTime": 1733053809557,
       "endTime": 0,
       "sourceId": "process-catalog",
       "sourceName": "process catalog",
@@ -16,7 +16,7 @@
     },
     {
       "identity": "process-catalog-to-process-cash-desk-2",
-      "startTime": 1674202485242353,
+      "startTime": 1733210658557,
       "endTime": 0,
       "sourceId": "process-catalog",
       "sourceName": "process catalog",
@@ -44,7 +44,7 @@
     },
     {
       "identity": "process-payment-1-to-process-cash-desk-1",
-      "startTime": 1674202485242353,
+      "startTime": 1732790614557,
       "endTime": 0,
       "sourceId": "process-payment-1",
       "sourceName": "process payment 1",
@@ -58,8 +58,8 @@
     },
     {
       "identity": "process-payment-2-to-process-catalog",
-      "startTime": 1674202485242353,
-      "endTime": 1674202485242354,
+      "startTime": 1732308411557,
+      "endTime": 1732816557557,
       "sourceId": "process-payment-2",
       "sourceName": "process payment 2",
       "destinationId": "process-catalog",
@@ -72,7 +72,7 @@
     },
     {
       "identity": "process-cart-1-to-process-payment-3",
-      "startTime": 1674202485242353,
+      "startTime": 1732842483557,
       "endTime": 0,
       "sourceId": "process-cart-1",
       "sourceName": "process cart 1",
@@ -86,7 +86,7 @@
     },
     {
       "identity": "process-database-1-to-process-payment-1",
-      "startTime": 1674202485242353,
+      "startTime": 1732566550557,
       "endTime": 0,
       "sourceId": "process-database-1",
       "sourceName": "process database 1",
@@ -100,7 +100,7 @@
     },
     {
       "identity": "site-client-1-to-process-database-1",
-      "startTime": 1674202485242353,
+      "startTime": 1732336898557,
       "endTime": 0,
       "sourceId": "site-client-1",
       "sourceName": "site client 1",
@@ -114,7 +114,7 @@
     },
     {
       "identity": "process-cash-desk-2-to-process-payment-1",
-      "startTime": 1674202485242353,
+      "startTime": 1732710931557,
       "endTime": 0,
       "sourceId": "process-cash-desk-2",
       "sourceName": "process cash desk 2",
@@ -128,7 +128,7 @@
     },
     {
       "identity": "process-catalog-to-process-cart-1-tcp",
-      "startTime": 1674202485242353,
+      "startTime": 1732957099557,
       "endTime": 0,
       "sourceId": "process-catalog",
       "sourceName": "process catalog",
